refactor: ♻️ featuresディレクトリのインポートをパブリックAPIに置き換えました
