+++ packages/app/src/features/game/index.ts
@@ -1,0 +1,4 @@
+export * from "./game.model";
+export * from "./hooks/useGame";
+export * from "./hooks/useResetGame";
+export * from "./hooks/useUpdateGame";
