+++ packages/app/src/features/scene/index.ts
@@ -1,0 +1,2 @@
+export * from "./scene.hook";
+export * from "./scene.model";
